// src/modules/type/type.enum.ts
export enum TypeEntityEnum {
  Contract = 'Contract',
}

export enum TypeEnum {
  Temp = 'temp',
  Payroll = 'payroll',
  StudentTemp = 'studentTemp',
  ShortTermContract = 'shortTermContract',
  LongTermContract = 'longTermContract',
  StudentPayroll = 'studentPayroll',
  TemporaryContract = 'temporaryContract',
  CivilContract = 'civilContract',
  B2B = 'b2b',
  EmploymentContract = 'employmentContract',
}